fixed next build not working by dynamically importing component and modules. moved libs to api path
