+++ pages/components/Loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function Loading({ className }) {
+  return (
+    <div className={className} style={{
+      backgroundColor: "#272822",
+    }}>
+    </div>
+  )
+}
