--- pages/lib/input.js
@@ -1,1 +1,0 @@
-var a=['\x74\x61\x72\x67\x65\x74','\x67\x72\x65\x65\x74','\x6c\x6f\x67','\x48\x65\x6c\x6c\x6f\x20','\x73\x65\x74\x54\x61\x72\x67\x65\x74','\x77\x6f\x72\x6c\x64','\x72\x65\x61\x64\x65\x72'];var b=function(c,d){c=c-0x0;var e=a[c];return e;};(function(){class c{constructor(d){this[b('0x0')]=d;}[b('0x1')](){console[b('0x2')](b('0x3')+this[b('0x0')]);}[b('0x4')](e){this[b('0x0')]=e;}}const f=new c(b('0x5'));f[b('0x1')]();f[b('0x4')](b('0x6'));f[b('0x1')]();}());
